Collapse SafePathsSecureStorage into BTSecureStorage #661
#### Why:
`SafePathsSecureStorage` was originally written as a superclass for `BTSecureStorage` and `GPSSecureStorage`. Since `GPSSecureStorage` has been removed, there is no need for a superclass, and  `SafePathsSecureStorage` can be collapsed into `BTSecureStorage`.

#### This commit:
This commit removes `SafePathsSecureStorage`
