+++ ios/BT/Storage/BTSecureStorage.swift
@@ -2,13 +2,16 @@ import ExposureNotification
 import Foundation
 import RealmSwift
 
-class BTSecureStorage: SafePathsSecureStorage {
+class BTSecureStorage {
 
   static let shared = BTSecureStorage(inMemory: false)
 
-  override var keychainIdentifier: String {
+  var keychainIdentifier: String {
     "\(Bundle.main.bundleIdentifier!).realm"
   }
+
+  let identifier: String
+  let inMemory: Bool
 
   private lazy var realmConfig: Realm.Configuration = {
     guard let realmConfig = getRealmConfig() else {
@@ -39,15 +42,80 @@ class BTSecureStorage: SafePathsSecureStorage {
 
   private let notificationCenter: NotificationCenter
 
+  private lazy var keychainTag = keychainIdentifier.data(using: String.Encoding.utf8, allowLossyConversion: false)!
+  private lazy var keychainAccessControl = SecAccessControlCreateWithFlags(nil, kSecAttrAccessibleAfterFirstUnlock, [], nil)!
+
   init(inMemory: Bool = false, notificationCenter: NotificationCenter = NotificationCenter.default) {
     self.notificationCenter = notificationCenter
-    super.init(inMemory: inMemory)
+    self.identifier = UUID().uuidString
+    self.inMemory = inMemory
     if !userStateExists {
       resetUserState({ _ in })
     }
   }
 
-  override func getRealmConfig() -> Realm.Configuration? {
+  private lazy var keychainQuery: [CFString: Any] = [
+    kSecClass: kSecClassKey,
+    kSecAttrApplicationTag: keychainTag,
+    kSecAttrKeySizeInBits: 512,
+    kSecReturnData: true,
+    kSecMatchLimit: kSecMatchLimitOne,
+  ]
+
+  private lazy var keychainUpdateQuery: [CFString: Any] = [
+    kSecClass: kSecClassKey,
+    kSecAttrApplicationTag: keychainTag,
+    kSecAttrKeySizeInBits: 512,
+  ]
+
+  final func getEncryptionKey() -> NSData? {
+    // First check in the keychain for an existing key
+    // To avoid Swift optimization bug, should use withUnsafeMutablePointer() function to retrieve the keychain item
+    // See also: http://stackoverflow.com/questions/24145838/querying-ios-keychain-using-swift/27721328#27721328
+    var keychainData: CFTypeRef?
+    var status = withUnsafeMutablePointer(to: &keychainData) {
+      SecItemCopyMatching(keychainQuery as CFDictionary, UnsafeMutablePointer($0))
+    }
+
+    if status == errSecSuccess, let keychainData = keychainData as? NSData {
+      // For backwards compatibility, ensure existing items have the correct access control
+      status = SecItemUpdate(keychainUpdateQuery as CFDictionary, [
+        kSecAttrAccessControl: keychainAccessControl,
+      ] as CFDictionary)
+      assert(status == errSecSuccess, "Failed to set access control")
+
+      return keychainData
+    }
+
+    // No pre-existing key from this application, so generate a new one
+    let keyData = NSMutableData(length: 64)!
+    status = SecRandomCopyBytes(kSecRandomDefault, 64, keyData.mutableBytes.bindMemory(to: UInt8.self, capacity: 64))
+
+    guard status == errSecSuccess else {
+      assertionFailure("Failed to get random bytes")
+      return nil
+    }
+
+    // Store the key in the keychain
+    let query: [CFString: Any] = [
+      kSecClass: kSecClassKey,
+      kSecAttrApplicationTag: keychainTag,
+      kSecAttrAccessControl: keychainAccessControl,
+      kSecAttrKeySizeInBits: 512 as AnyObject,
+      kSecValueData: keyData,
+    ]
+
+    status = SecItemAdd(query as CFDictionary, nil)
+
+    guard status == errSecSuccess else {
+      assertionFailure("Failed to insert the new key in the keychain")
+      return nil
+    }
+
+    return keyData
+  }
+
+  func getRealmConfig() -> Realm.Configuration? {
     if let key = getEncryptionKey() {
       if (inMemory) {
         return Realm.Configuration(inMemoryIdentifier: identifier, encryptionKey: key as Data, schemaVersion: 11,
